Add files via upload
Overall, building the FiveSecTimer was a great way to practice my JavaScript, HTML, and Bootstrap skills. By building a simple timer with a button that modifies the timer, I was able to gain a better understanding of these technologies and how they can be used together to create interactive web applications.